change FloServerPingData.lossRate type from integer to float to account for new data from mm

+++ W3C.Contracts/Matchmaking/Flo/FloServerPingData.cs
@@ -5,6 +5,6 @@ public class FloServerPingData
     public int nodeId { get; set; }
     public int currentPing { get; set; }
     public int avgPing { get; set; }
-    public int lossRate { get; set; }
+    public float lossRate { get; set; }
     public int pingFilter { get; set; }
 }
